--- backend/contract-service/src/main/java/digit/config/MainConfiguration.java
@@ -1,39 +1,0 @@
-package digit.config;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.egov.tracer.config.TracerConfiguration;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
-
-import javax.annotation.PostConstruct;
-import java.util.TimeZone;
-
-
-@Import({TracerConfiguration.class})
-public class MainConfiguration {
-
-    @Value("${app.timezone}")
-    private String timeZone;
-
-    @PostConstruct
-    public void initialize() {
-        TimeZone.setDefault(TimeZone.getTimeZone(timeZone));
-    }
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        return new ObjectMapper().disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES).setTimeZone(TimeZone.getTimeZone(timeZone));
-    }
-
-    @Bean
-    @Autowired
-    public MappingJackson2HttpMessageConverter jacksonConverter(ObjectMapper objectMapper) {
-        MappingJackson2HttpMessageConverter converter = new MappingJackson2HttpMessageConverter();
-        converter.setObjectMapper(objectMapper);
-        return converter;
-    }
-}
