--- backend/contract-service/src/main/java/digit/util/UserUtil.java
@@ -1,140 +1,0 @@
-package digit.util;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import digit.config.Configuration;
-import digit.models.coremodels.UserDetailResponse;
-import digit.models.coremodels.user.Role;
-import digit.models.coremodels.user.User;
-import digit.models.coremodels.user.enums.UserType;
-import digit.repository.ServiceRequestRepository;
-import org.egov.tracer.model.CustomException;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Collections;
-import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.List;
-
-@Component
-public class UserUtil {
-
-
-    private ObjectMapper mapper;
-
-    private ServiceRequestRepository serviceRequestRepository;
-
-    @Autowired
-    private Configuration configs;
-
-
-    @Autowired
-    public UserUtil(ObjectMapper mapper, ServiceRequestRepository serviceRequestRepository) {
-        this.mapper = mapper;
-        this.serviceRequestRepository = serviceRequestRepository;
-    }
-
-    /**
-     * Returns UserDetailResponse by calling user service with given uri and object
-     *
-     * @param userRequest Request object for user service
-     * @param uri         The address of the endpoint
-     * @return Response from user service as parsed as userDetailResponse
-     */
-
-    public UserDetailResponse userCall(Object userRequest, StringBuilder uri) {
-        String dobFormat = null;
-        if (uri.toString().contains(configs.getUserSearchEndpoint()) || uri.toString().contains(configs.getUserUpdateEndpoint()))
-            dobFormat = "yyyy-MM-dd";
-        else if (uri.toString().contains(configs.getUserCreateEndpoint()))
-            dobFormat = "dd/MM/yyyy";
-        try {
-            LinkedHashMap responseMap = (LinkedHashMap) serviceRequestRepository.fetchResult(uri, userRequest);
-            parseResponse(responseMap, dobFormat);
-            UserDetailResponse userDetailResponse = mapper.convertValue(responseMap, UserDetailResponse.class);
-            return userDetailResponse;
-        } catch (IllegalArgumentException e) {
-            throw new CustomException("IllegalArgumentException", "ObjectMapper not able to convertValue in userCall");
-        }
-    }
-
-
-    /**
-     * Parses date formats to long for all users in responseMap
-     *
-     * @param responseMap LinkedHashMap got from user api response
-     */
-
-    public void parseResponse(LinkedHashMap responseMap, String dobFormat) {
-        List<LinkedHashMap> users = (List<LinkedHashMap>) responseMap.get("user");
-        String format1 = "dd-MM-yyyy HH:mm:ss";
-        if (users != null) {
-            users.forEach(map -> {
-                        map.put("createdDate", dateTolong((String) map.get("createdDate"), format1));
-                        if ((String) map.get("lastModifiedDate") != null)
-                            map.put("lastModifiedDate", dateTolong((String) map.get("lastModifiedDate"), format1));
-                        if ((String) map.get("dob") != null)
-                            map.put("dob", dateTolong((String) map.get("dob"), dobFormat));
-                        if ((String) map.get("pwdExpiryDate") != null)
-                            map.put("pwdExpiryDate", dateTolong((String) map.get("pwdExpiryDate"), format1));
-                    }
-            );
-        }
-    }
-
-    /**
-     * Converts date to long
-     *
-     * @param date   date to be parsed
-     * @param format Format of the date
-     * @return Long value of date
-     */
-    private Long dateTolong(String date, String format) {
-        SimpleDateFormat f = new SimpleDateFormat(format);
-        Date d = null;
-        try {
-            d = f.parse(date);
-        } catch (ParseException e) {
-            throw new CustomException("INVALID_DATE_FORMAT", "Failed to parse date format in user");
-        }
-        return d.getTime();
-    }
-
-    /**
-     * enriches the userInfo with statelevel tenantId and other fields
-     * The function creates user with username as mobile number.
-     *
-     * @param mobileNumber
-     * @param tenantId
-     * @param userInfo
-     */
-    public void addUserDefaultFields(String mobileNumber, String tenantId, User userInfo, UserType userType) {
-        Role role = getCitizenRole(tenantId);
-        userInfo.setRoles(Collections.singleton(role));
-        userInfo.setType(userType);
-        userInfo.setUsername(mobileNumber);
-        userInfo.setTenantId(getStateLevelTenant(tenantId));
-        userInfo.setActive(true);
-    }
-
-    /**
-     * Returns role object for citizen
-     *
-     * @param tenantId
-     * @return
-     */
-    private Role getCitizenRole(String tenantId) {
-        Role role = Role.builder().build();
-        role.setCode("CITIZEN");
-        role.setName("Citizen");
-        role.setTenantId(getStateLevelTenant(tenantId));
-        return role;
-    }
-
-    public String getStateLevelTenant(String tenantId) {
-        return tenantId.split("\\.")[0];
-    }
-
-}
