--- backend/contract-service/src/main/java/digit/web/models/ContractResponse.java
@@ -1,47 +1,0 @@
-package digit.web.models;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.validation.annotation.Validated;
-
-import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * ContractResponse
- */
-@Validated
-@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2023-02-01T15:45:33.268+05:30")
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ContractResponse {
-    @JsonProperty("responseInfo")
-    @Valid
-    private ResponseInfo responseInfo = null;
-
-    @JsonProperty("contracts")
-    @Valid
-    private List<Contract> contracts = null;
-
-    @JsonProperty("pagination")
-    @Valid
-    private Pagination pagination = null;
-
-
-    public ContractResponse addContractsItem(Contract contractsItem) {
-        if (this.contracts == null) {
-            this.contracts = new ArrayList<>();
-        }
-        this.contracts.add(contractsItem);
-        return this;
-    }
-
-}
-
