+++ backend/contract-service/src/main/java/org/egov/works/Main.java
@@ -1,0 +1,20 @@
+package org.egov.works;
+
+
+import org.egov.tracer.config.TracerConfiguration;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Import;
+
+@Import({TracerConfiguration.class})
+@SpringBootApplication
+@ComponentScan(basePackages = {"digit", "digit.web.controllers", "digit.config"})
+public class Main {
+
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(Main.class, args);
+    }
+
+}
