+++ backend/contract-service/src/main/java/org/egov/works/kafka/Consumer.java
@@ -1,0 +1,20 @@
+package org.egov.works.kafka;
+
+import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
+
+@Component
+public class Consumer {
+
+    /*
+     * Uncomment the below line to start consuming record from kafka.topics.consumer
+     * Value of the variable kafka.topics.consumer should be overwritten in application.properties
+     */
+    //@KafkaListener(topics = {"kafka.topics.consumer"})
+    public void listen(final HashMap<String, Object> record) {
+
+        //TODO
+
+    }
+}
