+++ backend/contract-service/src/main/java/org/egov/works/web/models/ContractRequest.java
@@ -1,0 +1,38 @@
+package org.egov.works.web.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.egov.common.contract.request.RequestInfo;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.Valid;
+
+/**
+ * ContractRequest
+ */
+@Validated
+@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2023-02-01T15:45:33.268+05:30")
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ContractRequest {
+    @JsonProperty("RequestInfo")
+    @Valid
+    private RequestInfo requestInfo = null;
+
+    @JsonProperty("contract")
+    @Valid
+    private Contract contract = null;
+
+    @JsonProperty("workflow")
+    @Valid
+    private Workflow workflow = null;
+
+
+}
+
