+++ backend/contract-service/src/main/java/org/egov/works/web/models/Document.java
@@ -1,0 +1,44 @@
+package org.egov.works.web.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.swagger.annotations.ApiModel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.constraints.Size;
+
+/**
+ * This object holds list of documents attached during the transaciton for a property
+ */
+@ApiModel(description = "This object holds list of documents attached during the transaciton for a property")
+@Validated
+@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2023-02-01T15:45:33.268+05:30")
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Document {
+    @JsonProperty("id")
+    @Size(max = 64)
+    private String id = null;
+
+    @JsonProperty("documentType")
+    private String documentType = null;
+
+    @JsonProperty("fileStore")
+    private String fileStore = null;
+
+    @JsonProperty("documentUid")
+    @Size(max = 64)
+    private String documentUid = null;
+
+    @JsonProperty("additionalDetails")
+    private Object additionalDetails = null;
+
+
+}
+
