+++ backend/contract-service/src/main/java/org/egov/works/web/models/Workflow.java
@@ -1,0 +1,45 @@
+package org.egov.works.web.models;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.springframework.validation.annotation.Validated;
+
+import javax.validation.constraints.NotNull;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Workflow
+ */
+@Validated
+@javax.annotation.Generated(value = "org.egov.codegen.SpringBootCodegen", date = "2023-02-01T15:45:33.268+05:30")
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Workflow {
+    @JsonProperty("action")
+    @NotNull
+    private String action = null;
+
+    @JsonProperty("comment")
+    private String comment = null;
+
+    @JsonProperty("assignees")
+    private List<String> assignees = null;
+
+
+    public Workflow addAssigneesItem(String assigneesItem) {
+        if (this.assignees == null) {
+            this.assignees = new ArrayList<>();
+        }
+        this.assignees.add(assigneesItem);
+        return this;
+    }
+
+}
+
