--- backend/contract-service/src/test/java/digit/TestConfiguration.java
@@ -1,16 +1,0 @@
-package digit;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.KafkaTemplate;
-
-import static org.mockito.Mockito.mock;
-
-@Configuration
-public class TestConfiguration {
-    @Bean
-    @SuppressWarnings("unchecked")
-    public KafkaTemplate<String, Object> kafkaTemplate() {
-        return mock(KafkaTemplate.class);
-    }
-}
