+++ backend/contract-service/src/test/java/org/egov/works/TestConfiguration.java
@@ -1,0 +1,16 @@
+package org.egov.works;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.core.KafkaTemplate;
+
+import static org.mockito.Mockito.mock;
+
+@Configuration
+public class TestConfiguration {
+    @Bean
+    @SuppressWarnings("unchecked")
+    public KafkaTemplate<String, Object> kafkaTemplate() {
+        return mock(KafkaTemplate.class);
+    }
+}
